Refactor nav tag into a NavBar component
