+++ src/assets/NavBar.module.css
@@ -1,0 +1,61 @@
+:root {
+  --body-color: #14471e;
+  --text-color: #fff;
+  --hover-color: #ee9b01;
+}
+
+.wrapper__nav {
+  background-color: var(--body-color);
+  color: var(--text-color);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 5px 10px;
+}
+
+.nav__logo a {
+  text-decoration: none;
+  color: var(--text-color);
+  font-weight: 700;
+}
+
+.nav__menu {
+  display: flex;
+}
+
+.nav__menu li {
+  padding: 5px 30px;
+  cursor: pointer;
+}
+
+.nav__menu li:hover {
+  background-color: var(--hover-color);
+  border-radius: 2px;
+}
+
+.nav__user span {
+  cursor: pointer;
+}
+
+@media screen and (max-width: 700px) {
+  .wrapper__nav {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  .nav__menu {
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+  }
+
+  .nav__menu li {
+    width: 100%;
+    text-align: center;
+    padding: 5px 10px;
+  }
+
+  .wrapper__movies {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
