+++ src/router/Home.js
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
-import Movie from "../components/Movie.js";
 import { Link } from "react-router-dom";
 import styles from "../assets/Home.module.css";
+import Movie from "../components/Movie.js";
+import NavBar from "../components/NavBar.js";
 
 function Home() {
   const [loading, setLoading] = useState(true);
   const [movies, setMovies] = useState([]);
-  const movieGenres = ["Popular", "Action", "Animation", "Family", "Sci-Fi"];
 
   const getMovies = async () => {
     const res = await (
@@ -29,19 +29,7 @@ function Home() {
         <h3>Loading...</h3>
       ) : (
         <div className="">
-          <nav className={styles.wrapper__nav}>
-            <div className={styles.nav__logo}>
-              <Link to="/">MOVIE</Link>
-            </div>
-            <ul className={styles.nav__menu}>
-              {movieGenres.map((genre) => (
-                <li key={genre}>{genre}</li>
-              ))}
-            </ul>
-            <div className={styles.nav__user}>
-              <span>LOGIN</span>
-            </div>
-          </nav>
+          <NavBar />
           <section className={styles.wrapper__movies}>
             {movies.map((movie) => (
               <Link to={`/movie/${movie.id}`}>
